maintain native linux app compat for testing
