+++ lib/main.dart
@@ -3,10 +3,4 @@ import 'package:picotracker_client/pico_app.dart';
 
 void main() {
   runApp(const PicoApp());
-  try {
-    // SerialPort(linuxUsbTTY).close();
-  } catch (e) {
-    // ignore: avoid_print
-    print("could not close serial port $e");
-  }
 }
