+++ lib/main_screen.dart
@@ -6,11 +6,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:picotracker_client/command_builder.dart';
 import 'package:picotracker_client/picotracker/screen_char_grid.dart';
+import 'package:picotracker_client/serialportinterface.dart';
 
 import 'commands.dart';
 import 'widgets/pico_screen.dart';
-import 'package:webserial/webserial.dart';
-import 'dart:js_interop';
 
 class MainScreen extends StatefulWidget {
   const MainScreen({super.key, required this.title});
@@ -26,15 +25,16 @@ class _MainScreenState extends State<MainScreen> {
   int keymask = 0;
   StreamSubscription? subscription;
   StreamSubscription? cmdStreamSubscription;
-  // SerialPort? port;
   final _grid = ScreenCharGrid();
   final cmdBuilder = CmdBuilder();
+  late final SerialPortHandler serialHandler;
 
   StreamSubscription? usbUdevStream;
 
   @override
   void initState() {
     super.initState();
+    serialHandler = SerialPortHandler(cmdBuilder);
 
     cmdBuilder.commands.listen((cmd) {
       setState(() {
@@ -59,60 +59,6 @@ class _MainScreenState extends State<MainScreen> {
     subscription?.cancel();
     cmdStreamSubscription?.cancel();
     usbUdevStream?.cancel();
-  }
-
-  void chooseSerialDevice() async {
-    late final JSSerialPort? port;
-    try {
-      port = await requestWebSerialPort();
-      print("got serial port: $port");
-    } catch (e) {
-      print(e);
-      return;
-    }
-
-    if (port?.readable == null) {
-      // Open the serial port.
-      await port
-          ?.open(
-            JSSerialOptions(
-              baudRate: 115200,
-              dataBits: 8,
-              stopBits: 1,
-              parity: "none",
-              bufferSize: 64,
-              flowControl: "none",
-            ),
-          )
-          .toDart;
-
-      print("port opened: ${port?.readable}");
-    } else {
-      print("port already opened: ${port?.readable}");
-    }
-    // Listen to data coming from the serial device.
-    final reader = port?.readable?.getReader() as ReadableStreamDefaultReader;
-
-    while (true) {
-      final result = await reader.read().toDart;
-      if (result.done) {
-        // Allow the serial port to be closed later.
-        reader.releaseLock();
-        break;
-      } else {
-        // value is a Uint8Array.
-        onSerialData(result.value as JSUint8Array);
-      }
-    }
-  }
-
-  void onSerialData(JSUint8Array data) {
-    // print("$data");
-
-    final byteBuf = data.toDart;
-    for (int i = 0; i < byteBuf.lengthInBytes; i++) {
-      cmdBuilder.addByte(byteBuf[i]);
-    }
   }
 
   @override
@@ -164,7 +110,7 @@ class _MainScreenState extends State<MainScreen> {
           floatingActionButton: MaterialButton(
             child: const Text("choose serial port"),
             onPressed: () {
-              chooseSerialDevice();
+              serialHandler.chooseSerialDevice();
             },
           ),
           body: Stack(
