+++ lib/serialport_none.dart
@@ -1,0 +1,7 @@
+import 'command_builder.dart';
+
+class SerialPortHandler {
+  SerialPortHandler(CmdBuilder cmdBuilder);
+
+  void chooseSerialDevice() async {}
+}
