+++ lib/serialportinterface.dart
@@ -1,0 +1,3 @@
+export 'serialport_none.dart'
+    if (dart.library.io) 'serialport_native.dart'
+    if (dart.library.js_interop) 'serialport_web.dart';
