+++ pubspec.yaml
@@ -10,6 +10,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  flutter_libserialport: ^0.4.0
   webserial: ^1.0.0
 
 dev_dependencies:
